fix(57): file_size i32→i64 with BigIntExportBehavior::Number, add best-effort comment to interrupt_task

+++ src-tauri/src/models/task.rs
@@ -79,7 +79,7 @@ pub struct TaskAttachment {
     pub task_id: i32,
     pub filename: String,
     pub file_path: String,
-    pub file_size: i32,
+    pub file_size: i64,
     pub created_at: String,
 }
